# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于Gin + Vue + Element UI的前后端分离权限管理系统的接口文档\n添加qq群: 74520518 进入技术交流群 请备注，谢谢！",
        "title": "kubetabs API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://github.com/wenjianzhang/kubetabs/blob/master/LICENSE.md"
        },
        "version": "0.0.1"
    },
    "paths": {
        "/api/v1/config": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取JSON",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "配置"
                ],
                "summary": "修改配置",
                "parameters": [
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.Config"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": -1, \"message\": \"添加失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/config/{configId}": {
            "get": {
                "security": [
                    {
                        "": []
                    }
                ],
                "description": "获取JSON",
                "tags": [
                    "配置"
                ],
                "summary": "获取配置",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "配置编码",
                        "name": "configId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"data\": [...]}",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除数据",
                "tags": [
                    "配置"
                ],
                "summary": "删除配置",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "configId",
                        "name": "configId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": -1, \"message\": \"删除失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/configKey/{configKey}": {
            "get": {
                "security": [
                    {
                        "": []
                    }
                ],
                "description": "获取JSON",
                "tags": [
                    "配置"
                ],
                "summary": "获取配置",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "configKey",
                        "name": "configKey",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"data\": [...]}",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/configList": {
            "get": {
                "security": [
                    {
                        "": []
                    }
                ],
                "description": "获取JSON",
                "tags": [
                    "配置"
                ],
                "summary": "配置列表数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "configKey",
                        "name": "configKey",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "configName",
                        "name": "configName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "configType",
                        "name": "configType",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页条数",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "pageIndex",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"data\": [...]}",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/dict/data": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取JSON",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "字典数据"
                ],
                "summary": "修改字典数据",
                "parameters": [
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.DictType"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": -1, \"message\": \"添加失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取JSON",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "配置"
                ],
                "summary": "添加配置",
                "parameters": [
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.Config"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": -1, \"message\": \"添加失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/dict/data/list": {
            "get": {
                "security": [
                    {
                        "": []
                    }
                ],
                "description": "获取JSON",
                "tags": [
                    "字典数据"
                ],
                "summary": "字典数据列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "dictCode",
                        "name": "dictCode",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "dictType",
                        "name": "dictType",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页条数",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "pageIndex",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"data\": [...]}",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/dict/data/{dictCode}": {
            "get": {
                "security": [
                    {
                        "": []
                    }
                ],
                "description": "获取JSON",
                "tags": [
                    "字典数据"
                ],
                "summary": "通过编码获取字典数据",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "字典编码",
                        "name": "dictCode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"data\": [...]}",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除数据",
                "tags": [
                    "字典数据"
                ],
                "summary": "删除字典数据",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "dictCode",
                        "name": "dictCode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": -1, \"message\": \"删除失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/dict/databyType/{dictType}": {
            "get": {
                "security": [
                    {
                        "": []
                    }
                ],
                "description": "获取JSON",
                "tags": [
                    "字典数据"
                ],
                "summary": "通过字典类型获取字典数据",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "dictType",
                        "name": "dictType",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"data\": [...]}",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Config": {
            "type": "object",
            "properties": {
                "configId": {
                    "description": "编码",
                    "type": "integer"
                },
                "configKey": {
                    "description": "参数键名",
                    "type": "string"
                },
                "configName": {
                    "description": "参数名称",
                    "type": "string"
                },
                "configType": {
                    "description": "是否系统内置",
                    "type": "string"
                },
                "configValue": {
                    "description": "参数键值",
                    "type": "string"
                },
                "createBy": {
                    "type": "string"
                },
                "createTime": {
                    "type": "string"
                },
                "dataScope": {
                    "type": "string"
                },
                "isDel": {
                    "type": "string"
                },
                "params": {
                    "type": "string"
                },
                "remark": {
                    "description": "备注",
                    "type": "string"
                },
                "updateBy": {
                    "type": "string"
                },
                "updateTime": {
                    "type": "string"
                }
            }
        },
...
        "models.DictType": {
            "type": "object",
            "properties": {
                "createBy": {
                    "description": "创建者",
                    "type": "string"
                },
                "createTime": {
                    "description": "创建时间",
                    "type": "string"
                },
                "dataScope": {
                    "type": "string"
                },
                "dictId": {
                    "type": "integer",
                    "example": 1
                },
                "dictName": {
                    "description": "字典名称",
                    "type": "string"
                },
                "dictType": {
                    "description": "字典类型",
                    "type": "string"
                },
                "isDel": {
                    "type": "string"
                },
                "params": {
                    "type": "string"
                },
                "remark": {
                    "description": "备注",
                    "type": "string"
                },
                "status": {
                    "description": "状态",
                    "type": "string"
                },
                "updateBy": {
                    "description": "更新者",
                    "type": "string"
                },
                "updateTime": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
...
        "models.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "代码",
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "description": "数据集",
                    "type": "object"
                },
                "msg": {
                    "description": "错误消息",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}